Rename module interally to grass
