+++ .test.env
@@ -22,8 +22,9 @@ REDIS_PASSWORD=pass
 ####### Actinia ########
 ACTINIA_API_KEY=eyJhbGciOiJIUzUxMiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoiYWN0aW5pYS1nZGkifQ.UPeFRT3HA2MtKUQBJEgkeiEj5OxmODp8ZoAXWw0J776WRxeSOOCj17PtRnJiV6K8FfJb7zPS7frsZVCOv5trww
 ACTINIA_VERSION=v3
-ACTINIA_BASEURL=actinia-core:8088
+ACTINIA_BASEURL=localhost:8088
 ACTINIA_USER=actinia-gdi
 ACTINIA_PASSWORD=actinia-gdi
 ACTINIA_LOCATION=nc_spm_08_grass7
 ACTINIA_MAPSET=PERMANENT
+ACTINIA_SWAGGER_URL=http://localhost:8088/api/v3/swagger.json
