+++ grass/models/ObjectInfoAbstract.py
