+++ grass/models/SimpleResponseAbstract.py
@@ -5,7 +5,7 @@
 # Author: Corey White (smortopahri@gmail.com)                                  #
 # Maintainer: Corey White                                                      #
 # -----                                                                        #
-# Last Modified: Wed Oct 18 2023                                               #
+# Last Modified: Mon Nov 13 2023                                               #
 # Modified By: Corey White                                                     #
 # -----                                                                        #
 # License: GPLv3                                                               #
@@ -32,7 +32,7 @@
 
 from django.db import models
 
-from actinia.models.enums import ResponseStatusEnum
+from grass.models.enums import ResponseStatusEnum
 from .fields.ActiniaResourceStatusEnumField import ActiniaResourceStatusEnumField
 from .fields.ActiniaResponseStatusEnumField import ActiniaResponseStatusEnumField
