+++ setup.cfg
@@ -1,5 +1,5 @@
 [metadata]
-name = django-actinia
+name = django-grass
 version = 0.0.1-alpha.0
 description = A Django package to utilize Actinia REST API for GRASS GIS for distributed compuational analysis and modeling.
 long_description = file: README.rst
@@ -20,7 +20,7 @@ install_requires =
     Django >= 4.2
 
 [build_sphinx]
-project= 'django-actinia'
+project= 'django-grass'
 version=0.0.1-alpha.0
 release=0.0.1-alpha.0
 source-dir= 'docs'
