+++ tests/models/test_Region.py
@@ -5,7 +5,7 @@
 # Author: Corey White (smortopahri@gmail.com)                                  #
 # Maintainer: Corey White                                                      #
 # -----                                                                        #
-# Last Modified: Fri Nov 10 2023                                               #
+# Last Modified: Mon Nov 13 2023                                               #
 # Modified By: Corey White                                                     #
 # -----                                                                        #
 # License: GPLv3                                                               #
@@ -32,7 +32,7 @@
 
 
 from django.test import TestCase
-from actinia.models import Region
+from grass.models import Region
 
 
 class RegionModelTest(TestCase):
